@@ -1,18 +1,18 @@
 import 'package:flutter/material.dart';
-import 'package:simple_nft/pages/discover.dart';
 import 'package:simple_nft/pages/get_started.dart';
-import 'package:simple_nft/pages/signin.dart';
+import 'package:simple_nft/theme.dart';
 
 void main() => runApp(const MyApp());
 
 class MyApp extends StatelessWidget {
-  const MyApp({Key? key}) : super(key: key);
+  const MyApp({super.key});
 
-  @override 
+  @override
   Widget build(BuildContext context) {
-    return const MaterialApp(
+    return MaterialApp(
       debugShowCheckedModeBanner: false,
-      home: GetStarted(),
+      theme: appTheme(),
+      home: const GetStarted(),
     );
   }
 }
